fixed no new node bug
need to make a new node each time the insertAfter function is called - because if you use the same node then  you are having the code just change its next and prev node -- you end up splicing large portions of text at a time. which is great for splicing but bad for this.